+++ src/commands/rename.rs
@@ -387,7 +387,7 @@ mod tests {
         installed.packages.insert(
             "hello".to_string(),
             InstalledPackage {
-                meta_version: crate::core::manifest::CURRENT_META_VERSION,
+                schema_version: crate::core::manifest::CURRENT_SCHEMA_VERSION,
                 repo_name: "hello".to_string(),
                 variant: None,
                 version: "script".to_string(),
@@ -426,7 +426,7 @@ mod tests {
         exe1.insert("bin/oldcmd".to_string(), "oldcmd".to_string());
 
         let package = InstalledPackage {
-            meta_version: crate::core::manifest::CURRENT_META_VERSION,
+            schema_version: crate::core::manifest::CURRENT_SCHEMA_VERSION,
             repo_name: "pkg1".to_string(),
             variant: None,
             version: "1.0.0".to_string(),
@@ -458,7 +458,7 @@ mod tests {
         exe1.insert("bin/cmd1".to_string(), "cmd1".to_string());
 
         let package1 = InstalledPackage {
-            meta_version: crate::core::manifest::CURRENT_META_VERSION,
+            schema_version: crate::core::manifest::CURRENT_SCHEMA_VERSION,
             repo_name: "pkg1".to_string(),
             variant: None,
             version: "1.0.0".to_string(),
@@ -483,7 +483,7 @@ mod tests {
         exe2.insert("bin/cmd2".to_string(), "cmd2".to_string());
 
         let package2 = InstalledPackage {
-            meta_version: crate::core::manifest::CURRENT_META_VERSION,
+            schema_version: crate::core::manifest::CURRENT_SCHEMA_VERSION,
             repo_name: "pkg2".to_string(),
             variant: None,
             version: "1.0.0".to_string(),
@@ -562,7 +562,7 @@ mod tests {
         let mut exe_cli = HashMap::new();
         exe_cli.insert("bin/confy-64".to_string(), "confy-64".to_string());
         let pkg_cli = InstalledPackage {
-            meta_version: crate::core::manifest::CURRENT_META_VERSION,
+            schema_version: crate::core::manifest::CURRENT_SCHEMA_VERSION,
             repo_name: "confy".to_string(),
             variant: Some("64".to_string()),
             version: "0.19.1".to_string(),
@@ -585,7 +585,7 @@ mod tests {
         let mut exe_desktop = HashMap::new();
         exe_desktop.insert("bin/confyd".to_string(), "confyd".to_string());
         let pkg_desktop = InstalledPackage {
-            meta_version: crate::core::manifest::CURRENT_META_VERSION,
+            schema_version: crate::core::manifest::CURRENT_SCHEMA_VERSION,
             repo_name: "confy".to_string(),
             variant: Some("desktop-64".to_string()),
             version: "0.19.1".to_string(),